+++ R/apache-httpd.R
@@ -10,20 +10,32 @@
 #' @export
 apache_httpd_version_history <- function() {
 
-  readr::read_lines("https://archive.apache.org/dist/httpd/") %>%
-    purrr::keep(stri_detect_regex, 'apache_.*gz"') %>%
-    stri_replace_first_fixed("<img src=\"/icons/compressed.gif\" alt=\"[   ]\"> ", "") %>%
-    stri_match_first_regex('href="apache_([[:digit:]]+\\.[[:digit:]]+\\.[[:digit:]]+)\\.tar\\.gz".*([[:digit:]]{4}-[[:digit:]]{2}-[[:digit:]]{2})') %>%
-    dplyr::as_data_frame() %>%
-    dplyr::select(-V1) %>%
-    dplyr::rename(vers = V2, rls_date = V3) %>%
-    dplyr::mutate(rls_date = as.Date(rls_date)) %>%
-    dplyr::mutate(rls_year = lubridate::year(rls_date)) %>%
+  ap <- readr::read_lines("https://archive.apache.org/dist/httpd/")
+
+  apd <- xml2::read_html(paste0(ap[grepl('"(httpd-|apache_)[[:digit:]]+\\.[[:digit:]]+\\.[[:digit:]]+\\.tar', ap)], collapse="\n"))
+
+  rvest::html_text(apd) %>%
+    stri_split_lines() %>%
+    unlist() %>%
+    stri_match_first_regex("([[:digit:]]{4}-[[:digit:]]{2}-[[:digit:]]{2})") %>%
+    .[,2] -> rls_dates
+
+  rvest::html_nodes(apd, "a") %>%
+    rvest::html_attr("href") %>%
+    stri_match_first_regex("[-_]([[:digit:]]+\\.[[:digit:]]+\\.[[:digit:]]+)\\.tar") %>%
+    .[,2] -> vers
+
+  dplyr::tibble(
+    vers = vers,
+    rls_date = rls_dates
+  ) %>%
+    dplyr::distinct(vers, .keep_all=TRUE) %>%
+    mutate(rls_date = as.Date(rls_date)) %>%
+    mutate(rls_year = lubridate::year(rls_date)) %>%
     dplyr::bind_cols(
       semver::parse_version(.$vers) %>%
-        dplyr::as_data_frame()
+        dplyr::as_tibble()
     ) %>%
     dplyr::arrange(major, minor, patch) %>%
     dplyr::mutate(vers = factor(vers, levels = vers))
-
 }