+++ R/chrome.R
@@ -29,7 +29,7 @@ google_chrome_version_history <- function() {
     dplyr::distinct(vers, .keep_all=TRUE) %>%
     dplyr::bind_cols(
       semver::parse_version(.$vers) %>%
-        dplyr::as_data_frame()
+        dplyr::as_tibble()
     ) %>%
     dplyr::arrange(major, minor, patch) %>%
     dplyr::mutate(vers = factor(vers, levels = vers))